+++ src/main/java/kitchenpos/ordertable/domain/NumberOfGuests.java
@@ -3,7 +3,7 @@ package kitchenpos.ordertable.domain;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
-import kitchenpos.common.Value;
+import kitchenpos.common.domain.Value;
 
 @Embeddable
 public class NumberOfGuests extends Value<NumberOfGuests> {
