+++ src/main/java/kitchenpos/product/domain/ProductName.java
@@ -3,7 +3,7 @@ package kitchenpos.product.domain;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
-import kitchenpos.common.Value;
+import kitchenpos.common.domain.Value;
 
 @Embeddable
 public class ProductName extends Value<ProductName> {
