+++ src/test/java/kitchenpos/common/domain/PriceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.common;
+package kitchenpos.common.domain;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.ThrowableAssert.*;
@@ -10,6 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import kitchenpos.common.domain.Price;
 
 @DisplayName("가격")
 class PriceTest {
