+++ src/test/java/kitchenpos/common/domain/QuantityTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.common;
+package kitchenpos.common.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import kitchenpos.common.domain.Quantity;
 
 @DisplayName("수량")
 class QuantityTest {
