+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import kitchenpos.common.Price;
+import kitchenpos.common.domain.Price;
 
 @DisplayName("상품")
 class ProductTest {
